feat(h5): 支持 chooseMedia 方法
